Complete intial refactoring of BattleEntity, BattleCog
Dataclasses and inheritance are a pain in the butt to debug.
Additionally, I've ran into many circular imports due to my desire to type hint,
so I've moved the BattleStates into Battle.py.

Next step:
- Create functionality for BattleCog's attacks and targets.
- Write tests to verify a BattleCog cannot attack or target another BattleCog.

partial fix for #87
