+++ Cog.py
@@ -9,24 +9,19 @@ from .CogGlobals import COG_ATTRIBUTES, get_cog_vitals
 from .Entity import Entity
 
 
-@dataclass
+@dataclass(init=False, repr=False)
 class Cog(Entity):
     key: str
     relative_level: Optional[int] = field(default=0)
 
-    # Prevent needing to pass in Attack's required arguments
-    # Instead, pass in the required arguments during __post_init__()
-    name: str = field(init=False)
-    hp: int = field(init=False)
-    # # For testing purposes. See `test_cog_attack_damages_multiple_toons`
-
-    def __post_init__(self):
+    def __init__(self, key: str, relative_level: int):
+        self.key = key
+        self.relative_level = relative_level
         self.vitals = get_cog_vitals(cog_key=self.key,
                                      relative_level=self.relative_level)
-        self.name = self.vitals['name']
         self.hp_max = self.vitals['hp']
+        super().__init__(name=self.vitals['name'], hp=self.vitals['hp'])
 
-        super().__init__(name=self.name, hp=self.hp_max)
         self.attacks = self.vitals['attacks']
         self.defense = self.vitals['def']
         self.level = self.vitals['level']
