+++ tests/battle/test_battle_entity_creation.py
@@ -1,15 +1,12 @@
 
-from ...Battle import BattleEntity
+from ...Entity import BattleEntity
 
 
 class TestEntityCreation:
 
-    name = "Test"
-    hp = 10
     battle_id = 1
 
-    def test_entity_creation(self):
-        e = BattleEntity(name=self.name, hp=self.hp, battle_id=self.battle_id)
-        assert e.name == self.name
-        assert e.hp == self.hp
-        assert e.battle_id == self.battle_id
+    def test_battle_entity_creation(self):
+        be = BattleEntity(battle_id=self.battle_id)
+        assert be.battle_id == self.battle_id
+        assert isinstance(be, BattleEntity)
